PCIeBinding: Fix display of BDF in device register log (#79)
The BDF in the "device registered" log of PCIe binding is displayed as
garbled characters since std::hex will not be applied to char (the
underlying type of uint8_t). Fix this issue by casting it to int first.

Tested:
Verified BDF is printed as hex in log when device is registered.

+++ src/PCIeBinding.cpp
@@ -437,11 +437,13 @@ void PCIeBinding::processRoutingTableChanges(
              */
             std::stringstream busHex, deviceHex, functionHex;
             busHex << std::setfill('0') << std::setw(2) << std::hex
-                   << hw::bdf::getBus(pciePrivate->remote_id);
+                   << static_cast<int>(hw::bdf::getBus(pciePrivate->remote_id));
             deviceHex << std::setfill('0') << std::setw(2) << std::hex
-                      << hw::bdf::getDevice(pciePrivate->remote_id);
+                      << static_cast<int>(
+                             hw::bdf::getDevice(pciePrivate->remote_id));
             functionHex << std::hex
-                        << hw::bdf::getFunction(pciePrivate->remote_id);
+                        << static_cast<int>(
+                               hw::bdf::getFunction(pciePrivate->remote_id));
 
             std::string bus(busHex.str()), device(deviceHex.str()),
                 function(functionHex.str());
